+++ Modules/Intent.Modules.AspNetCore.Controllers/release-notes.md
@@ -5,7 +5,7 @@
 ### Version 4.0.2
 
 - Fixed: Errors thrown when applying `application/json` to the Return Type Mediatype the Http Settings of an operation.
-- Explicity adds `application/json` mediatype explicitly to the controller operation OpenAPI metadata attributes.
+- Adds `[Produces(MediaTypeNames.Application.Json)]` when explicitly specified in the Http Settings.
 
 ### Version 4.0.1